requirements.txt
streamlit
surya-ocr